+++ src/main/kotlin/com/lykke/matching/engine/services/CashSwapOperationService.kt
@@ -17,6 +17,7 @@ import com.lykke.matching.engine.outgoing.messages.JsonSerializable
 import com.lykke.matching.engine.round
 import com.lykke.matching.engine.utils.RoundingUtils
 import org.apache.log4j.Logger
+import java.math.BigDecimal
 import java.util.Date
 import java.util.LinkedList
 import java.util.UUID
@@ -42,7 +43,7 @@ class CashSwapOperationService(private val balancesHolder: BalancesHolder,
 
         val balance1 = balancesHolder.getBalance(message.clientId1, message.assetId1)
         val reservedBalance1 = balancesHolder.getReservedBalance(message.clientId1, message.assetId1)
-        if (balance1 - reservedBalance1 < operation.volume1) {
+        if (balance1 - reservedBalance1 < BigDecimal.valueOf(operation.volume1)) {
             messageWrapper.writeNewResponse(ProtocolMessages.NewResponse.newBuilder().setId(message.id).setMatchingEngineId(operation.id)
                     .setStatus(LOW_BALANCE.type).setStatusReason("ClientId:${message.clientId1},asset:${message.assetId1}, volume:${message.volume1}").build())
             LOGGER.info("Cash swap operation failed due to low balance: ${operation.clientId1}, ${operation.volume1} ${operation.asset1}")
@@ -51,7 +52,7 @@ class CashSwapOperationService(private val balancesHolder: BalancesHolder,
 
         val balance2 = balancesHolder.getBalance(message.clientId2, message.assetId2)
         val reservedBalance2 = balancesHolder.getReservedBalance(message.clientId2, message.assetId2)
-        if (balance2 - reservedBalance2 < operation.volume1) {
+        if (balance2 - reservedBalance2 < BigDecimal.valueOf(operation.volume1)) {
             messageWrapper.writeNewResponse(ProtocolMessages.NewResponse.newBuilder().setId(message.id).setMatchingEngineId(operation.id)
                     .setStatus(LOW_BALANCE.type).setStatusReason("ClientId:${message.clientId2},asset:${message.assetId2}, volume:${message.volume2}").build())
             LOGGER.info("Cash swap operation failed due to low balance: ${operation.clientId2}, ${operation.volume2} ${operation.asset2}")
@@ -84,14 +85,14 @@ class CashSwapOperationService(private val balancesHolder: BalancesHolder,
         val operations = LinkedList<WalletOperation>()
 
         operations.add(WalletOperation(UUID.randomUUID().toString(), operation.externalId, operation.clientId1, operation.asset1,
-                operation.dateTime, -operation.volume1))
+                operation.dateTime, BigDecimal.valueOf(-operation.volume1)))
         operations.add(WalletOperation(UUID.randomUUID().toString(), operation.externalId, operation.clientId2, operation.asset1,
-                operation.dateTime, operation.volume1))
+                operation.dateTime, BigDecimal.valueOf(operation.volume1)))
 
         operations.add(WalletOperation(UUID.randomUUID().toString(), operation.externalId, operation.clientId1, operation.asset2,
-                operation.dateTime, operation.volume2))
+                operation.dateTime, BigDecimal.valueOf(operation.volume2)))
         operations.add(WalletOperation(UUID.randomUUID().toString(), operation.externalId, operation.clientId2, operation.asset2,
-                operation.dateTime, -operation.volume2))
+                operation.dateTime, BigDecimal.valueOf(-operation.volume2)))
 
         balancesHolder.createWalletProcessor(LOGGER).preProcess(operations).apply(operation.externalId, MessageType.CASH_SWAP_OPERATION.name)
     }
